backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
aiofiles==23.2.1
torch==2.3.0
torchvision==0.18.0
timm==1.0.3
Pillow==10.3.0
pandas==2.2.2
scikit-learn==1.5.0
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2